Put Cheep and Author into their own files
